REFACTOR login routers.
DONE:
1. Build a general login middleware for both endpoints.
2. Refactor all files accordingly.
